scmi_perf/test: Add FAST_CHANNELS test target
This patch adds a new test target which has the
BUILD_HAS_SCMI_PERF_FAST_CHANNELS feature flag enabled. An existing
unit test has been updated to make use of this.

In an upcoming patch, unit tests for the function
`scmi_perf_describe_fast_channels()` will be added, which requires
this feature flag to be enabled.

Co-authored-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Signed-off-by: Chuyue Luo <Chuyue.Luo@arm.com>
Change-Id: Ia1c8a0df0a0702c921f577e9ce475aaa5b3f5c34