smcf: Low level register map and hardware control
The System Monitoring Control Framework is designed to manage a large and
diverse set of on-chip sensors and monitors. It does this by presenting
software with a standard interface to control the monitors, regardless of
type, and reducing software load of controlling the monitor sampling and data
collection.

The SMCF reduces the burden on monitor control by enabling sampling on multiple
monitors to be controlled together and by various triggers either internal or
external to the SMCF. The number of monitors that the SMCF supports can be
configured. The SMCF eases data collection requirements by allowing the data
from multiple monitors to be collated in a single location or writing out data
to a memory-mapped location that is easier for the monitoring agent to access.

The SMCF can also reduce the requirement on the monitoring agent to constantly
monitor data by providing programmable alerts that can inform the monitoring
agent when certain changes happen, or thresholds are crossed.

This patch adds the initial support for SMCF. The patch provides register map
and low level hardware functionality internal to this module.
The next patch will provide the main module functionality and the interface
to the rest of the firmware.

Signed-off-by: Tarek El-Sherbiny <tarek.el-sherbiny@arm.com>
Change-Id: I98c6c5699c31aef1c91a7940e4e7c0e7bf697f73