tcop: Evaluate perf limits at start
Traffic-Cop evaluates the performance limits upon the cores being
turned ON and/or OFF. When the platforms boots, the primary core is
already ON and the limits evaluation is not performed.
This leaves the internal data uninitialised (zeroed) and in particular
the performance limit that TCOP would return to the plugins handler.
This would cause the platform to boot at the minimum performance level
until the next core within the same group is turned ON.

Add evaluation of the performance limits at start such that the
internal data is properly initialised and consistent with the initial
power status of the cores.

Change-Id: Iae3308e919b5eaf4d6956915a218c214dafec65c
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>