mod/mhu3: Fix doorbell channel trigger interrupt
To trigger an interrupt on a doorbell channel
corresponding PDBCW_ST bit of the PBX(postbox) is validated.
If corresponding bit is unset then the code will raise
the doorbell interrupt, otherwise it will return an error
FWK_E_STATE. This validation check is incorrect and this
change fixes this.

Change-Id: I84b19832b1d128176a2b77f2f5abc3c3cb1d92c0
Signed-off-by: Girish Pathak <girish.pathak@arm.com>