product/rdn2: Move from module smt to transport
This change makes use of a more generic transport module instead of
smt for the rdn2 platform. The new transport module has additional
features such as in-band message transport and fast channel
interface that are missing in the traditional smt.

Since the transport module provides all the features of the smt
module, the smt module will be removed entirely from the codebase
after all other dependent platforms moves to use the transport module.

Change-Id: Iffcf13d3ed159bb5dd6182b511729c91722a9e88
Signed-off-by: Girish Pathak <girish.pathak@arm.com>