module/mhu: Enable code to use transport module.
The transport module is a replacement for the the smt module.
However, until all platforms migrate to the transport,
the mhu module conditionally(build time) must support
both smt as well as transport.

This change will enable module/mhu to
use either smt or transport based on the build
configuration.

Change-Id: I6140e72845d88853c14cde56635023257da6a6d7
Signed-off-by: Girish Pathak <girish.pathak@arm.com>