module/smt: Remove smt module.
The smt module is replaced by the transport module.
Remove it from codebase.

The new transport module has additional features such as
in-band message transport and fast channel interface that
are missing in the traditional smt.

Since the transport module provides all the features of the smt
module, the smt module is removed entirely from the codebase.

Change-Id: Ib3a161179f9789a9731d1df4880262431c595024
Signed-off-by: Girish Pathak <girish.pathak@arm.com>