product/rdn2: reduce use of PLATFORM_VARIANT build option
The core count is one of the significant difference between the various
RD-N2 platform variants. The NUMBER_OF_CLUSTERS macro defines the number
of core/cluster for a variant. In preparation to add another variant of
RD-N2 platform, replace the use of PLATFORM_VARIANT build flag, where
applicable, with the NUMBER_OF_CLUSTERS macro. This helps to reduce the
changes required to add support for a new variant.

Signed-off-by: Tony K Nadackal <tony.nadackal@arm.com>
Change-Id: Id1f9ddd2032384b3f8a1e679b0b191ae258e34b1