thermal_mgmt: adds activity factor API driver
This patch adds activity factor API driver functionality that allows
to accumulate idle power from each actor during operation and distribute
it amongst other actors to use it. Activity factor is an optional
feature that can be configured when it is required.

The driver API is platform specific because it gives flexibility in
terms of implementation depending on system configuration.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I4407e41865d2ebcd01c894eeaaa72ecf16596eb5