unit_test: Add 'ignore' in the config and recompile mockers
With unit testing, There are some cases where we are not interested
about the number of times a particular function is called.
In those cases the `ignore` feature of CMock comes handy.

Add such support in the config and re-run all the mockers so that
they contain the ignore mockers as well.

Change-Id: I2f6464b451d6bb4c62b64b86509c72c4911bfbb7
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>