scmi_perf: add logs when multiple fast channels events are queued
This patch adds a log message when multiple fast channels events
are pending to be processed. This will allow to identify issues
related with the processing time of performance plugins when they are
configured.
It is also increased the log level from `TRACE` to `ERROR` when
it is not possible to create a fast channels process event.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I154735381342a363c101209122c15c5e562983b0