mhu3: Correct the conditional check for MHU in/out channel
MHU can be implemented unidirectional as well, like the fastchannel
implementation between application processor and xCP. In this case, it
is not mandatory to have both input and output channels. For the case of
AP-xCP fastchannel, from xCP's perspective, the MHU is unidirectional,
and there is no output channel required. Update the conditional check in
MHU initialization sequence to return error only when both in and out
channels are absent.

Signed-off-by: Pranav Madhu <pranav.madhu@arm.com>
Signed-off-by: Thomas Abraham <thomas.abraham@arm.com>
Change-Id: Ib7628cd3b4a3322486b10110ab0b7c9ca6126113