cmn700: move counters and flags to the context
Move the IO and Sys Cache group regions counters and scg_region_enabled
flags from cmn700_setup_sam to the module context.

This is an incremental change to add APIs to add and remove memory
regions in RNSAM. These APIs are required to enable the support to
update/add regions in CMN700 during runtime.

Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Change-Id: Icda38a9a1e1028b61fd113faecbcf1e45bd3adfb