fwk_io: remove prepend carriage return from module
Currently, the terminator string is partially handled by fwk_log by
adding a \n character at the end of each log message when it is needed.
Later, when the log message is processed by the driver, it prepends a
carriage return character (\r) on specified conditions.

By doing this, the responsibility of setting the end line is delegated to
the framework and the driver itself, which is incorrect and breaks
encapsulation, and makes it hardware-agnostic.

For this mentioned reason, this responsibility is being taken down from
the driver. In an upcoming patch, the log terminator string is going to be
refactored to improve configurability and encapsulation.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I45be77bb6cec2296e1176bbb422c1fdcac939707