pl011: add `FWK_E_BUSY` when TX buffer is full
Previously, PL011 driver waits until there was a free space in
the TX buffer. However this was a blocking loop controlled by
external hardware which is a software vulnerabilty.

This was modified to make the driver return `FWK_E_BUSY` when
TX buffer is full and delegate the resonsability to handle
this to the framework.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: Ie8e09c0589711da4603704393684f0b50436abb5