fwk_core: Fix buffered log output behaviour
When the subsystem mode was disabled, the system would be suspended
even if there were characters waiting to be printed. This resulted
in only one character being printed each time the system woke up.
The issue was resolved by preventing the system from being suspended
when there were elements in the output log buffer that were ready
to be printed.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I63322343130b84ba8f10294340cf64d413c07931