fwk_log: refactor banner and log terminator configuration
The log terminator was fixed and not possible to be changed/configured,
delegating the responsibility to modify it to the `stream` driver.

This patch includes a firmware definition `FMW_LOG_ENDLINE_STR` that allows
to define a custom log terminator if desired.

It is also refactored the banner to be aligned to the log terminator
modifications.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: Ia896a8ee2772e5a5c1fc83e95ddfea7db61c8b47