docker: fix dependencies and broken links
This patch fixes all outdated dependencies and broken links.

ArmClang compiler 6 is not longer provided in the Docker image,
if it is required the compiler directory should be mounted as a
volumen in the specified location in the documentation.

LLVM was updated to the latest supported version by the firmware
with the instalation script that build the compiler-rt assets.

For the `dev` taget an user was created following the recomendations
from Docker documentation.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I28bc940fefa754796e98f473bdb1f85bd13cc683