module/atu: add driver module for address translation unit
The Address Translation Unit(ATU) provides means to convert the logical
address (32 bit LA) from the MSCP memory space to the physical address
residing in the system memory.

Add a driver module for the ATU that provides an interface for modules
to request a translation region to be mapped, unmapped or output bus
attributes to be configured during the runtime phase.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: Ic9bf58cf75c74ecc2d760e041369f3df824e1efa