module/system_power: fix to correct omitted notifications
Currently when an SCMI message to set the state is sent to the
`mod_scmi_system_power` module, if the state sent is one of
shutdown, warm reset or cold reset, a notification will not be
sent to subscribers.

For the other cases, if a state transition was successful it will
send a notification before going to exit. However, in this case
regardless of whether or not it was successful, it will go straight
to the exit section and miss sending a notification.

This has been amended to now only proceed to the exit if it is an
unsuccessful change instead, and will therefore include the call
to notify subscribers if the state change was executed as expected.

Signed-off-by: Katherine Vincent <katherine.vincent@arm.com>
Change-Id: I9c7b2602b2fd316408daec627abbf7781c5825e2