module/system_power: refactoring of the set_state function
Currently `scmi_sys_power_state_set_handler` is a fairly
long function with multiple cases for various states and
different actions and checks for each state.

To simplify this and increase readability, this is a
change to split some of the functionality up.

The fundamental functionality of `scmi_sys_power_state_set_handler`
should not be changing, just rearranging to make it easier
to maintain.

Signed-off-by: Katherine Vincent <katherine.vincent@arm.com>
Change-Id: I6afd1820b498b74ec4b48572fab55ef431053c08