rdn2/ramfw: Rearrange the order of module initialization
Rearrange the order of module initialization for RD-N2 platform in such
a way that the power related modules are initialized before the clock
and other dependent modules. This ensures that power domain is enabled
before clocks are enabled. This allows reducing inter-module dependency
using notifications and allows these hardware elements to be enabled by
default during boot.

Signed-off-by: Pranav Madhu <pranav.madhu@arm.com>
Change-Id: I53d7fab3dff4cb1b6112ddf7de4256193e4f184b