Fixes multiple cases of explicit type casting unsigned int(uint32_t) to int
As unsigned integers can store larger values than the maximum value that can be stored in a signed integer, hence in explicitly type cast from an unsigned int to an int, there is a potential issue of loss of information which leads to unexpected result. Also when the most significant bit  is set in unsigned int, it will be considered as a sign bit in int which eventually leads to wrong result.

So a good wrap-around is to make the data type of variable  “unsigned int” instead of “int”, which will not consume any extra memory compared to int and also give full range support of unsigned int values.

Change-Id: Id686c7b662bc5a0d03c0a47d1fa71d89438bc642
Signed-off-by: Sanchit Kumar <sanchit.dabgotra@gmail.com>