module/clock: allow initial rate to be set during initialization
The current clock implementation relies on power domain notification to
setup the clocks. In other words, the clocks are turned ON only after
receiving the power ON notification from the power domain which the
clocks are reliant on. If the power domain is enabled by default on
early boot phase, then the clock is not required to wait for a power
domain transition notification.

Add a new config data field named 'default_on' that can be used
to specify whether the clock can be setup with power domain transition
notification. If this flag is set, the clock is setup during early
initialization only if the corresponding power domain is found to be
enabled.

Change-Id: Ia7a22d076dc2bd1f9009db35f14dda34d4ac9f08
Signed-off-by: Pranav Madhu <pranav.madhu@arm.com>
Signed-off-by: Girish Pathak <girish.pathak@arm.com>