product/rdn2: Remove power domain id from transport module config
The power domain required for the transport and MHU module are
guaranteed to be enabled early in the boot. The transport module now
does not need to wait for the power domain transition notification to do
the start runtine. Hence remove the pd_source_id field from transport
module configuration file.

Change-Id: Ib0ed9ca13a13a645382b1606d893b81dbde60480
Signed-off-by: Pranav Madhu <pranav.madhu@arm.com>