smcf data: Add data handling functions
This commit adds a new layer called smcf_data in a separate file. The layer
provides two main features. The first feature provides all related functions
to unpack, extract and copy the data. The second feature provides the data
validation functions based on the sample identifiers provided, if any.
Because all sample identifiers are optional, the patch adds functions to check
if each identifier field is available in the sample header before performing
the validation. In addition, to support data read from RAM and as well as from
the MGI registers, pointers are used. The pointers are set at initialization to
point at MGI or RAM according to the data configuration passed in.

Signed-off-by: Tarek El-Sherbiny <tarek.el-sherbiny@arm.com>
Co-authored-by: Abhishek Rawat <abhishek.rawat@arm.com>
Change-Id: I76cac513a3dd96edfa779eb913c8b635ab6cf0d6