product/synquacer: prevent metadata from being updated on every boot
Current Trial Boot implementation for SynQuacer platform
updates metadata on every boot even if the value is not changed.
The metadata is stored in the NOR Flash, so updating NOR Flash
should be minimized.

This commit checks the accept flag, then updates the metadata
when the value is changed and needs to be updated.

Acked-by: Jassi Brar <jaswinder.singh@linaro.org>
Signed-off-by: Masahisa Kojima <masahisa.kojima@linaro.org>
Change-Id: I4bacc393f9cfb43e390d7958e052c25a59d8a8d0