docker: Install dependencies on image build
This patch modifies Dockerfile to install the python and ruby
dependencies specified in requirements.txt and Gemfile on image build.
The script `/usr/local/bin/dependencies` is removed as a result.

This changes the command needed to build the docker image. The relevant
places where `docker build` is used are updated accordingly.

Signed-off-by: Yat Long Poon <yatlong.poon@arm.com>
Change-Id: I8475dafb8ac03657043955dbb29a06e66796a616