fwk_io: Replace `fwk_assert` for error handling
`fwk_assert` calls  were removed to minimize assertions in the
framework code, since they can be easily replaced by error
handling. The reasons to do that are that the framework should be
reliable enough, the debugging process of the framework should not rely
on the users and the error handling can facilitate the usage of the
framework.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: Ib9c71580cb0c489b50a63d9fb5957ecf9459d212