mod/power_allocator: add mod_power_allocator.h
module/power_allocator is a place holder for the entity
that will do the power management in the SCP and with which
the SCMI Power Capping module will interact.

The header file is added in order to identify the interface provided
to SCMI Power Capping and complete SCMI Power Capping unit tests.

Signed-off-by: Ahmed Gadallah <ahmed.gadallah@arm.com>
Change-Id: I88b5be8217dc5605980d07d4103f848b6f22af7f