module/cdns_i2c: Add bus active check to APIs
When multiple API requests are made in succession, a new transfer could
be triggered while the I2C bus was still active. This request and any
subsequent requests would then fail until the system is shutdown.

A check is added to prevent the API from triggering new activity on the
I2C bus until the Bus Active field is clear, allowing for successive
API requests.

Signed-off-by: Werner Lewis <werner.lewis@arm.com>
Change-Id: I1f4919762229bce47e6a23aadeed1dce5ee303c9