module: Introduce sc_pll driver module
The PLLs in both the Morello and N1SDP platforms are Silicon
Creations-provided IPs and are present as product-specific
modules within the each product.

Generalise the Silicon Creations PLL driver by making it a
common SCP driver module. The private PLL module(s) and
references to them will be removed from the N1SDP and
Morello products in the imminent commits.

Signed-off-by: Vishnu Satheesh <vishnu.satheesh@arm.com>
Change-Id: Icec8342036f02f424b809e50faa20aaf0534ad6c