mod/power_meter: add mod_power_meter.h
module/power_meter is a place holder for the entity
that will do the power measurements in the SCP and with
which the SCMI Power Capping module will interact.

The header file is added in order to identify the interface provided
to SCMI Power Capping and complete SCMI Power Capping unit tests.

Signed-off-by: Ahmed Gadallah <ahmed.gadallah@arm.com>
Change-Id: I35660d3f3cf9c717c078718325959a6cf7228859