interface/amu: add AMU interface definition
Add definition of AMU interface to be implemented by multiple
drivers as counters access from SCP differs across platforms.

The interface introduces an API to get the AMU counters of CPUs.

The user of this API is required to provide a buffer to hold the
value of the counters as well as the `counter_id` to start from
and the number of counters needed.

Signed-off-by: Mohamed Omar Asaker <mohamed.omarasaker@arm.com>
Change-Id: I9d4b7855c5152f5e5ca73a97d40204c49b5e734f