mod/power_coordinator: add mod_power_coordinator.h
module/power_coordinator is a place holder for the entity that will
coordinate different power management components' timing in the SCP
and with which the SCMI Power Capping module will interact.

The header file is added in order to identify the interface provided
to SCMI Power Capping and complete SCMI Power Capping unit tests.

Signed-off-by: Ahmed Gadallah <ahmed.gadallah@arm.com>
Change-Id: I75c888d274755fdf6c5313c9c6f17bcba4487bce