SMCF: Fix enablement of START and END sample identifiers in WRCFG
This change fixes the enablement of start and end sample identifiers.
Currently if start sample identifier is already enabled and then end
identifer is also requested, it results in WRCFG.NUM_SAMPLE_ID to be
writen with an invalid value 0b11 whereas the expected value as per
specification is 0b10 to enable both START and END sample identifiers
in data header.

Signed-off-by: Abhishek Rawat <abhishek.rawat@arm.com>
Change-Id: Iabaac4ef21700faf397a2e4aa3130f0baddc8fef