rdfremont: add config data for transport module in scpfw
Three transport channels are required for supporting platform boot -
channel for communicating PSCI messages with the application core,
channel for requesting the secure platform agent firmware to configure
the ATU and a channel for various handshake with the secure platform
agent firmware during boot. Add these as configuration data for the
transport module.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: Ib72034eae8ee388bd97ef70731fe7e6617d7d20f