rdfremont: add sid module config data for scp_ramfw
The SID peripheral provides information about the platform and its
configuration. So, add the configuration data for sid module in scp
ramfw. The configuration data includes the expected value of its
PID/CID registers for identification. PID5, PID6 and PID7 are not
implemented by this peripheral and this is indicated using the valid
register bitmap.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I9ce8ad9eff33dedf32a6129fc5181871fe595fc3