tc1/ram_fw: Fix inclusion of resource permissions in target
The module resource_perms is not actually appended in the list of
SCP_MODULES in the CMake files. That means that even with
SCP_ENABLE_RESOURCE_PERMISSIONS_INIT set to true, the resource
permission feature is not available in the ram_fw target.

Fix the above by adding "resource-perms" in the list of SCP_MODULES
in the correct CMake file.

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: I4422ecefe68652438404ba71b7e3afcf84f5f70e