apcontext: allow optional platform notification
On platforms that require additional boot time setup (in addition to the
clock configuration) to access the AP context memory region, the
platform notification can be subscribed. An example use of platform
notification is when another compute agent in the system has to setup
access permissions for the AP context memory region. This is optional
and if provided as module configuration data, the AP context memory
region will be accessed only after this notification is processed

Signed-off-by: Thomas Abraham <thomas.abraham@arm.com>
Change-Id: I3e5a3bbe08a75feb584b032f5dbb1e3119526169