sds: allow optional platform notification
On platforms that require additional boot time setup (in addition to the
clock configuration) to access the SDS memory regions, the platform
notification can be subscribed. An example use of platform notification
is when another compute agent in the system has to setup access
permissions for the SDS memory regions. This is optional and if provided
as module configuration data, the SDS memory regions will be accessed
only after this notification is processed.

Signed-off-by: Thomas Abraham <thomas.abraham@arm.com>
Change-Id: I830e48ef571081af6038cc401b04c490af2b5f46