transport: allow optional platform notification
On platforms that require additional boot time setup (in addition to the
power domain configuration) to access the mailbox memory regions, the
platform notification can be subscribed. An example use of platform
notification is when another compute agent in the system has to setup
access  permissions for the mailbox memory regions. This is optional and
if provided as module configuration data, the mailbox memory regions
will be accessed only after this notification is processed.

Signed-off-by: Thomas Abraham <thomas.abraham@arm.com>
Change-Id: Ifca6089b683f9dc33914f06e261eeb936a621f87