interface/cmn: add a generic interface for cmn drivers
CMN module driver can change in between platforms, this makes it
difficult to write a generic module that uses CMN APIs to map regions
during runtime. This issue can be solved by providing a generic API
interface that CMN modules use to define it's API's and use the module
id provided by the platform config to bind to the correct CMN module.

Add CMN interface header in the common interface.

Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Change-Id: Idfcd9c8c2319c358877e5c26a2730f6b17363685