module/cmn_cyprus: assign RAIDs to requesting agents
In multichip platforms, when a remote request is routed to the local CCG
RA node, the LDID of the request node is converted to a globally unique
RAID using the LDID-to-RAID LUT register before being routed to the
remote CCG HA. So, calculate expanded RAIDs for local request agents and
configure them in CCG RA node.

Note: The expanded RAIDs are calculated with the assumption that the
mesh on the remote chip is symmetrical with same number of requesting
agents. With this assumption, the expanded RAID assignment will not
overlap with remote chip RAIDs ensuring the uniqueness.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I45dc4e62660e95337cd8d516c3bafcf7402230ea