module/cmn_cyprus: enable HN-S to track remote caching agents
All the transactions reaching the HN-S contain the LDID of the caching
agent which has sent the request. The HN-S node maintains the list of
LDID to RN-F Node ID mappings for snoop filter (SF) tracking in the
HNS_RN_CLUSTER_PHYSID register.

In multichip mode, the remote chip caching agents (remote RN-Fs, LCN)
can send request to the local HN-S. So, the HN-S node requires a valid
LDID to Node ID mapping to track the remote caching agents. The CCG HA
is a proxy for all remote caching agents. Hence, program the LDID to
Node ID lookup with the CCG HA node id and mark them as remote.
Currently, only the non-clustered mode is supported.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I4bb38f791ff5f5c01b2c1491d53925508396c6bf