module/cmn_cyprus: add support for CCLA direct connect mode
Add support for enabling CCLA direct connect mode for CML SMP
configurations. This mode enables direct connection of the CXS interface
from the CCLA on one CMN‑Cyprus to the CXS interface of the other. In
this mode, these interfaces can be connected without going through the
lower link layer and PHY controller IP.

Note: This mode is used during pre-silicon validation stage and not used
when external controller IP is present.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I0ebfc700fa3e719c2fd3b11ce61bcf396a78ee1e