neoverse-rd/interface: add a generic interface for io block
Neoverse reference design platforms include an IO block that connects
the PCIe bus or other SoC-specific IO devices to the compute subsystem.
On the RD-N2 platform, NIC does not support the dynamic configuration of
the system address map and PCIe integration controller is used to
support it. For subsequent generations of platforms, the IO block has
been generalized by using a programmable NIC and SMMU to interface with
PCIe or other IO devices to the compute subsystem.

As there are now two variants of the IO block supporting the Neoverse
reference design platforms, the common PCIe bus configuration drivers
will need a platform-independent abstraction to interface with them.
Therefore, a new interface definition is required to provide a
platform-independent abstraction for common code to interface with the
various IO block variants.

Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Change-Id: Ica7d7ed1c352d6ebed3107e0635a546274e54829