module/cmn_cyprus: enable CML protocol link
In a multichip configuration, a protocol link can be established between
the CML on the local chip and a corresponding remote CML link that the
gateway can communicate with. So, add support to enable the CCG link
and request link up with remote protocol link as a preliminary step to
bringing up a CML system.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: If027bd0e8040e92bb8d09a20ca374aaa0d2f3153