module/cmn_cyprus: enable CCGs to enter coherency domain
CMN‑Cyprus includes a feature that allows RNs to be included or excluded
from the system coherency domain. This includes Snoop and DMV domain and
this feature ensures correct operations of Snoops and DVMs. Each CML
gateway block has software-programmable bits to allow the protocol links
to enter and exit the system coherency domains and DVM domains.

So, enable the CML protocol links to enter snoop coherency domain and
DVM domain in multichip configurations.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: Iad39876efdd5ae762040494db5fa742e3dfb4556