clk: Guard clock module event ids under BUILD_HAS_MOD_CLOCK
This change adds guards for clock module events identifiers in order
to make it optional for platforms that don't have a dependency
on it. This will help to remove clock module from platforms where
it is defined as empty config just for compilation. It also enable
the flag for DVFS unit test.

Signed-off-by: Abhishek Rawat <abhishek.rawat@arm.com>
Change-Id: I022e8a0b25daf944d77185c5854363372754d95e