rdfremont: add config data for mhuv3 driver in scpfw
Two instances of the MHUv3 doorbell channel are being initially added as
configuration data for the MHUv3 module. The first doorbell channel is
used to send messages between AP and SCP while the second doorbell
channel used between RSS and SCP. The configuration data for these
channels include the interrupt number and the base addresses of the
corresponding Postbox and Mailbox.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I08e95b4bc7b19f1843bbb5b5b1e51e25972ab034