Add a GitHub Actions to auto comment PR
The SCP repo in GitHub has been deprecated. The repo is now located at
GitLab. An automatic message will be triggered using GitHub Actions to
comment on raised PRs to notify the author that contributions are no
longer accepted in GitHub and they have to submit a Merge Request in
GitLab.

Signed-off-by: Wen Ping Teh <wenping.teh@arm.com>
Change-Id: I92733a10e77b540c136e3d66eeb1d6385214f9a2