module/cmn_cyprus: fix incorrect format specifier for clang
The Clang compiler considers uint32_t data type as 'unsigned int'
whereas the GCC considers it as 'long unsigned int'. A potential
solution is to use PRIx format specifiers provided in the inttypes
header but it is not compatible with aarch64 host machines.

Hence, as a workaround, cast the uint32_t variables as unsigned int data
type when printing as the concerned variables are guaranteed to be under
16 bits.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I548f2709e2dfe460c176455b02982b33d1a2041e